Processed shop and OLC.
NOTE: If we start encountering strange bugs with object strings being not what they should be, it's likely caused by these changes. Part of the code seems to lean heavily on re-using object strings from the prototypes (all pointers point there unless changed), so it's possible I've swept up one of these pointers in the replacement and scheduled the proto for wiping.