bandaid
The optimal thing to do right now is:
Pick an archetype.
Get a datajack.
Do decking.

This is a bandaid, ugly patch to make that not necessarily the wisest idea, since deckers seem to be able to makes substantially more than any other archtype - and will continue to do so after this change.

Figured I may as well put something in here.