Update archetypes.cpp
the race of the adept archetype is ork, and the improved quickness power pushes them beyond the maximum, making it cost one per point (meaning they've spent 8 power points in CC).

Improved body does not have this issue because it's not pushing above that. I've simply replaced improved quickness with rapid healing, which is extremely useful for new players who tend to take damage anyway.