Reshuffling adept powers
Moves spell shroud and true sight to the NERP trainer. While they are functionally coded spells, as far as I can tell illusion spells and detection spells are self-buffs only, making them functionally useless in code.

The other part is adding a second adept trainer that moves the - quite frankly, really terrible abilities that are overpriced or completely outclassed by abilities that cost exactly the same - and putting a warning sign on them in an adjacent room.

It's not a no, but it's definitely a "Please be experienced before walking into this archetype." I'm still waffling on how necessary it is.