Adds additional toggle args
The description when using toggle without args versus the helpfile and the command to toggle a few things did not line up. This has been corrected.

Todo: Remove noradio, no repeat, and RPE unless you wish to cannibalize one of them (such as re-enabling RPE for questors, and so on).