Ugly as sin, but it compiles
Moves has_key from do_move to utils.cpp and utils.hpp
Vehicles now straight-up ignore doors if there's a valid apartment they are a guest our own and also the driver has a key in their pocket.
Drones with MOD_POKEYSTICK and people on bikes can now push elevator buttons (there are no pokeystick mods yet).

None of this is properly tested, but it compiles without breaking. The logic for ignoring doors is also ugly as sin, but it'll probably work.