Vehicle decay (#650)
* First pass

In dblist.cpp, the logic for decaying vehicle needs to be moved to where the handler is for timers on characters (wherever that is).

Also probably needs a swath of typos and minor fixes, but it's getting there.

* debugging

* Update memory.cpp

* Stumped

* cleanup

* Update fight.cpp

* testing required

* Update comm.cpp

* update src\comm.cpp

 Immediately segfaults on use.

* Core logic rework

* updoots

* Update comm.cpp

* working on review

* review changes

* Update act.drive.cpp

* Update act.drive.cpp

segfaults. problem lies somewhere in spare, for and if loop works fine.

* testing

* at least it stopped crashing

* Update act.drive.cpp

* Update fight.cpp

* works now

* Fin

Thank you for all the help, Lucien.