Autodismissal of spirits at sunset (#670)
* Autodismissal of spirits at sunset

Previously, autodismissal of spirits at sunrise was changed to decrement a counter, with autodismissal when that counter runs out. This fixes sunset to do the same thing.

* Double the spirit lifetime counter

Because the spirit lifetime counter is decremented at both sunrise and sunset, it needs to be set to 2x NUMBER_OF_IG_DAYS_FOR_SPIRIT_TO_LAST on summoning.