Allow login->delete of hardcore characters (#676)
There already exists a check for dead hardcore characters when selecting the "Enter the game" option from the login menu. On the death of a hardcore character, the player does reach this menu and so has the option to delete the character.

Removing the additional check at the login prompt allows the player to decide to delete the character at a later time instead of only immediately after death.