Changes to initiative handling (#682)
* Fix NPCs attacking w/ negative initiative

NPCs are probably not flagged PRF_SEE_TIPS.

* Let new combat start w/ global init re-roll

Because of the global combat list, new combatants require re-ordering the list and thus can causes characters to experience arbitrary length out-of-initiative states throughout the duration of their fight. This prevents that, at the cost of delaying only the first round of new combats.

Also:
- No fight if no victim.
- Surprise flag set in new fights instead of with initiative rolls.
- Set init to 0 if surprised or not awake.

* Clarify out-of-init message

* Propagate combat list handling to set_fighting(ch, veh)

* Revert "Let new combat start w/ global init re-roll"

This reverts commit 56ed8a3f616f4d0951eaca60ac7f729c661cdff5.

* Revise new combatant initiative handling

The global initiative re-roll happens when the head of the combat list reaches 0 init. This prevents new combatants from replacing the head in order to bandage the problem of arbitrarily delays before the next global init re-roll.

* Revert "Clarify out-of-init message"

This reverts commit c45d1614de0cd41f7befa7163bd43200cecf32cb.