Calc Highest Bal/Imp separately
Currently a 4/2 and a 2/4 will come to a different total of armor (either 5/4 or 4/5)  depending on how they're being layered (either at the time of equipping, or based on the order in armor slots). This change should 'normalize' the formula, giving consistent results (5/5). If the same formula is used elsewhere and needs changed, I couldn't find it.