Initiative revision, limiting int version (#685)
* Initiative revision, limiting int version

This solution to the problem of arbitrary time until next global initiative reroll uses a limiting int, initiative_until_global_reroll. Its value is set by the initiative of the first combatant added to an empty combat_list, or by the highest initiative of a global init reroll. It is then reduced by 10 at the completion of each pass, and causes a global init reroll when <= 0.

* Decrement init tracker before while

While loop contains a return condition, so decrementing after the while will never happen.