Prevent SAN re-encryption when decker present
When attempting to logon to an adjacent host, the System Access Node (SAN) may be encrypted, and may be decrypted by targeting the exit's name.  In the mud, this is the same as the destination host's access subsystem being encrypted. In order to prevent a SAN from being repeatedly re-encrypted when a decker is attempting to pass through, we need to check for the presence of deckers in hosts surrounding the destination host.