Various decking fixes (#689)
* Limit hacking pool max by computer skill

The maximum number of Hacking Pool dice that can be added to any test is equal to the base number of skill dice in use. (SR3, pg 44).

* Prevent SAN re-encryption when decker present

When attempting to logon to an adjacent host, the System Access Node (SAN) may be encrypted, and may be decrypted by targeting the exit's name.  In the mud, this is the same as the destination host's access subsystem being encrypted. In order to prevent a SAN from being repeatedly re-encrypted when a decker is attempting to pass through, we need to check for the presence of deckers in hosts surrounding the destination host.

* Add messaging to 80 tally dump

Added feedback to player character about why they were dumped. Also added comment that not shutting down the host is a house rule.

* Fix truncation of matrix download percentage.

Previously, matrix download percentage was truncating to zero.

* Clarify programming step in comp list

Previously, software on a computer, when listed, showed (Design) regardless of whether it was being designed or programmed.

* Fix skill used to cap max hacking

* Clean up re-encryption checks

1. Put back checking same host for deckers
2. Integrate decker break condition in for loop control.

* Closed for-loop.

Co-authored-by: luciensadi <luciensadi@gmail.com>